refactor(defrag): split remaining cli store board and test hotspots (mem-e1097dd)
