+++ internal/store/store_gate_retry_collision_test.go
@@ -1,0 +1,108 @@
+package store
+
+import (
+	"context"
+	"strings"
+	"testing"
+)
+
+func TestGateCommandsRejectCommandIDsUsedByOtherEventTypes(t *testing.T) {
+	t.Parallel()
+
+	ctx := context.Background()
+	s := newTestStore(t)
+
+	if _, _, _, err := s.CreateIssue(ctx, CreateIssueParams{
+		IssueID:   "mem-a4b5c6d",
+		Type:      "task",
+		Title:     "Agent collision issue",
+		Actor:     "agent-1",
+		CommandID: "cmd-gate-collision-agent-1",
+	}); err != nil {
+		t.Fatalf("create agent collision issue: %v", err)
+	}
+	if _, _, _, err := s.CreateIssue(ctx, CreateIssueParams{
+		IssueID:   "mem-b5c6d7e",
+		Type:      "task",
+		Title:     "Human collision issue",
+		Actor:     "human:alice",
+		CommandID: "cmd-gate-collision-human-1",
+	}); err != nil {
+		t.Fatalf("create human collision issue: %v", err)
+	}
+
+	agentCases := []struct {
+		name string
+		call func() error
+	}{
+		{
+			name: "create gate template",
+			call: func() error {
+				_, _, err := s.CreateGateTemplate(ctx, CreateGateTemplateParams{
+					TemplateID:     "collision-template",
+					Version:        1,
+					AppliesTo:      []string{"task"},
+					DefinitionJSON: `{"gates":[]}`,
+					Actor:          "agent-1",
+					CommandID:      "cmd-gate-collision-agent-1",
+				})
+				return err
+			},
+		},
+		{
+			name: "instantiate gate set",
+			call: func() error {
+				_, _, err := s.InstantiateGateSet(ctx, InstantiateGateSetParams{
+					IssueID:      "mem-a4b5c6d",
+					TemplateRefs: []string{"collision-template@1"},
+					Actor:        "agent-1",
+					CommandID:    "cmd-gate-collision-agent-1",
+				})
+				return err
+			},
+		},
+		{
+			name: "lock gate set",
+			call: func() error {
+				_, _, err := s.LockGateSet(ctx, LockGateSetParams{
+					IssueID:   "mem-a4b5c6d",
+					Actor:     "agent-1",
+					CommandID: "cmd-gate-collision-agent-1",
+				})
+				return err
+			},
+		},
+		{
+			name: "evaluate gate",
+			call: func() error {
+				_, _, _, err := s.EvaluateGate(ctx, EvaluateGateParams{
+					IssueID:      "mem-a4b5c6d",
+					GateID:       "build",
+					Result:       "fail",
+					EvidenceRefs: []string{"ci://run/42"},
+					Actor:        "agent-1",
+					CommandID:    "cmd-gate-collision-agent-1",
+				})
+				return err
+			},
+		},
+	}
+
+	for _, tc := range agentCases {
+		t.Run(tc.name, func(t *testing.T) {
+			err := tc.call()
+			if err == nil || !strings.Contains(err.Error(), `command id already used by "issue.created"`) {
+				t.Fatalf("expected command collision error, got %v", err)
+			}
+		})
+	}
+
+	if _, _, err := s.ApproveGateTemplate(ctx, ApproveGateTemplateParams{
+		TemplateID: "missing-template",
+		Version:    1,
+		Actor:      "human:alice",
+		CommandID:  "cmd-gate-collision-human-1",
+	}); err == nil || !strings.Contains(err.Error(), `command id already used by "issue.created"`) {
+		t.Fatalf("expected approval command collision error, got %v", err)
+	}
+}
